# api/openapi.yaml
openapi: 3.1.2
info:
  title: Descendence API
  description: Basic api provided by the descendence api server
  version: v0.0.1
servers:
- url: http://127.0.0.1:8080
  description: Local development server for testing
paths:
...
  /api/v1/runs/{id}:
    get:
      operationId: getRun
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No run with this id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /api/v1/runtimes:
    get:
      operationId: listRuntimes
      description: >
        Runtimes, keyset-paginated by name - a catalogue, like jobs, not a
        timeline.
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Runtimes ordered by name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeList"
        "400":
          description: Malformed cursor, or limit outside 1-200
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      operationId: createRuntime
      description: >
        Define a runtime. Unlike a job, a runtime is not authored in git - it
        is a built artifact, created and rebuilt directly through this API
        (ARCHITECTURE.md §4.4).


        Leaving the row's build status at its default of `pending` queues the
        first build immediately; the supervisor's build claim loop picks it
        up on its next tick. Poll GET /api/v1/runtimes/{id} for buildStatus.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeCreate"
      responses:
        "202":
          description: Runtime created, its first build already queued
          headers:
            Location:
              description: "/api/v1/runtimes/{id} of the created runtime"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runtime"
        "400":
          description: Malformed body, missing/invalid lang, or an unrenderable definition
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: A runtime with this name already exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/runtimes/{id}:
    get:
      operationId: getRuntime
      description: >
        Returns the runtime. Also the poll target for an async build (task
        4.5): buildStatus and buildError live directly on this row, so there
        is no separate build resource.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The runtime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runtime"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No runtime with this id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
      required:
        - type
        - title
        - status
...
    Run:
      type: object
      properties:
        id:
          type: integer
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled, lost]
        imageRef:
          type: string
        argv:
          type: array
          items:
            type: string
        timeoutSeconds:
          type: integer
        containerId:
          type: [string, "null"]
        exitCode:
          type: [integer, "null"]
        failureReason:
          type: [string, "null"]
        cancelRequestedAt:
          description: >
            When cancellation was requested. Set before the run is actually
            cancelled - a running run's container still has to be stopped by
            the supervisor - so a run can be `running` with this set, meaning
            "on its way out". Also kept on a run that finished on its own
            before the request landed, which is the case worth being able to
            explain afterwards.
          type: [string, "null"]
          format: date-time
        queuedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
        jobId:
          description: >
            The job this run came from, or null for an ad-hoc run created by
            POST /api/v1/runs. A job may since have been removed from its
            repository; the run still points at it, so that what ran remains
            explainable.
          type: [integer, "null"]
        commitSha:
          description: >
            The exact commit the job's manifest and script were read from,
            resolved when the run was created and never re-resolved. Together
            with jobId this is what makes a past run reproducible: check this
            commit out and you have precisely what executed. Null for an
            ad-hoc run.
          type: [string, "null"]
        runtimeId:
          description: >
            Set when the job named a runtime rather than an image directly
            (task 4.6). Resolved once, at creation, from the runtime's
            current image digest - rebuilding the runtime afterwards never
            changes this run's imageDigest.
          type: [integer, "null"]
        imageDigest:
          description: The runtime's image digest at the moment this run was created. Null for an ad-hoc run or a job that names an image directly.
          type: [string, "null"]
      required:
        - id
        - state
        - imageRef
        - argv
        - timeoutSeconds
        - queuedAt
...
    Runtime:
      type: object
      description: >
        A built container image: curated base image, system packages, and
        one language's dependency manifest (ARCHITECTURE.md §4.4). Unlike a
        job, this is not authored in git - every field is owned by Postgres
        directly and changed only through this API.
      properties:
        id:
          type: integer
        name:
          description: Unique. How a job manifest's `runtime:` key addresses this row.
          type: string
        baseImage:
          type: string
        sysPackages:
          type: array
          items:
            type: string
        lang:
          type: string
          enum: [python, powershell, node]
        langManifest:
          description: >
            The language ecosystem's own manifest format verbatim -
            requirements.txt for python, a PSResourceGet file for
            powershell, package.json for node. Null means no language
            dependencies, just the base image and system packages.
          type: [string, "null"]
        inputHash:
          description: >
            Hash of (baseImage, sysPackages, lang, langManifest). Doubles as
            the local image tag, so two runtimes with identical inputs
            dedupe onto the same built image even under different names.
          type: string
        buildStatus:
          type: string
          enum: [pending, building, ready, failed]
        imageDigest:
          description: Set once buildStatus reaches ready.
          type: [string, "null"]
        buildError:
          description: Set when buildStatus is failed.
          type: [string, "null"]
        builtAt:
          type: [string, "null"]
          format: date-time
        imagePruned:
          description: >
            True once the image has been reclaimed (task 4.7). The row
            survives regardless - what was built, from which inputs, when,
            stays answerable - but a job naming this runtime cannot run
            until it is rebuilt.
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - baseImage
        - sysPackages
        - lang
        - inputHash
        - buildStatus
        - imagePruned
        - createdAt
    RuntimeList:
      type: object
      properties:
        items:
          description: Runtimes ordered by name.
          type: array
          items:
            $ref: "#/components/schemas/Runtime"
        nextCursor:
          description: Opaque cursor for the next page. Absent (or null) on the last page.
          type: [string, "null"]
      required:
        - items
    RuntimeCreate:
      type: object
      properties:
        name:
          type: string
        baseImage:
          description: Optional; defaults to a curated image for `lang` (decision #25, Debian-based).
          type: string
        sysPackages:
          type: array
          items:
            type: string
        lang:
          type: string
          enum: [python, powershell, node]
        langManifest:
          description: See Runtime.langManifest. Optional.
          type: string
      required:
        - name
        - lang
...